+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\sanyk\\Downloads\\Telegram Desktop\\googleMapApi\\google-api\\src\\index.js":"1","C:\\Users\\sanyk\\Downloads\\Telegram Desktop\\googleMapApi\\google-api\\src\\App.js":"2"},{"size":500,"mtime":499162500000,"results":"3","hashOfConfig":"4"},{"size":2874,"mtime":1606503044411,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1b7eps6",{"filePath":"8","messages":"9","errorCount":0,"warningCount":4,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"C:\\Users\\sanyk\\Downloads\\Telegram Desktop\\googleMapApi\\google-api\\src\\index.js",[],"C:\\Users\\sanyk\\Downloads\\Telegram Desktop\\googleMapApi\\google-api\\src\\App.js",["10","11","12","13"],{"ruleId":"14","severity":1,"message":"15","line":5,"column":16,"nodeType":"16","messageId":"17","endLine":5,"endColumn":25},{"ruleId":"14","severity":1,"message":"18","line":10,"column":3,"nodeType":"16","messageId":"17","endLine":10,"endColumn":15},{"ruleId":"14","severity":1,"message":"19","line":30,"column":5,"nodeType":"16","messageId":"17","endLine":30,"endColumn":10},{"ruleId":"14","severity":1,"message":"20","line":34,"column":5,"nodeType":"16","messageId":"17","endLine":34,"endColumn":21},"no-unused-vars","'getLatLng' is defined but never used.","Identifier","unusedVar","'ComboboxList' is defined but never used.","'ready' is assigned a value but never used.","'clearSuggestions' is assigned a value but never used."]
+[{"D:\\Project\\react-study-app\\src\\App.js":"1","D:\\Project\\react-study-app\\src\\reportWebVitals.js":"2","D:\\Project\\react-study-app\\src\\сomponents\\Map.js":"3","D:\\Project\\react-study-app\\src\\сomponents\\Event.js":"4","D:\\Project\\react-study-app\\src\\сomponents\\Card.js":"5","D:\\Project\\react-study-app\\src\\index.js":"6"},{"size":565,"mtime":1607565053088,"results":"7","hashOfConfig":"8"},{"size":375,"mtime":1607447722781,"results":"9","hashOfConfig":"8"},{"size":3248,"mtime":1607714875761,"results":"10","hashOfConfig":"8"},{"size":1433,"mtime":1607599789187,"results":"11","hashOfConfig":"8"},{"size":120,"mtime":1607566905247,"results":"12","hashOfConfig":"8"},{"size":517,"mtime":1607447722780,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"168iclj",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\Project\\react-study-app\\src\\App.js",[],"D:\\Project\\react-study-app\\src\\reportWebVitals.js",[],"D:\\Project\\react-study-app\\src\\сomponents\\Map.js",["26","27"],"D:\\Project\\react-study-app\\src\\сomponents\\Event.js",[],"D:\\Project\\react-study-app\\src\\сomponents\\Card.js",[],"D:\\Project\\react-study-app\\src\\index.js",[],{"ruleId":"28","severity":1,"message":"29","line":25,"column":7,"nodeType":"30","messageId":"31","endLine":25,"endColumn":12},{"ruleId":"28","severity":1,"message":"32","line":29,"column":7,"nodeType":"30","messageId":"31","endLine":29,"endColumn":23},"no-unused-vars","'ready' is assigned a value but never used.","Identifier","unusedVar","'clearSuggestions' is assigned a value but never used."]
