+++ package.json
@@ -3,12 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@reach/combobox": "^0.12.1",
+    "@react-google-maps/api": "^1.13.0",
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.2.2",
+    "carouselbox": "^1.2.3",
+    "date-fns": "^2.16.1",
     "react": "^17.0.1",
+    "react-bootstrap": "^1.4.0",
     "react-dom": "^17.0.1",
+    "react-elastic-carousel": "^0.9.5",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
+    "styled-components": "^5.2.1",
+    "use-places-autocomplete": "^1.6.0",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
