+++ src/сomponents/Card.css
@@ -1,0 +1,17 @@
+.card{
+   
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin: 0 100px;
+    width: 1000px;
+    height: 300px;
+    border-radius: 20px;
+    border: 10px solid white;
+    text-align: center;
+    font-size: 30px;
+    background: #333;
+    color: #fff;
+    
+
+}
