+++ src/сomponents/Card.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+
+const Card = ({number}) => <div className ="card">{number}</div>
+
+export default Card;
