+++ src/сomponents/Event.css
@@ -1,0 +1,51 @@
+.mr-auto{
+    display:flex;
+    font-size:50px;
+    
+
+    width: 20%;
+    height: 10%;
+    list-style-type:none;
+    margin: auto;
+    padding: 0px;
+    background-color: black;
+    justify-content: center;
+    
+}
+
+.text{
+    position: relative;
+    font-size: 20px;
+    margin: 50px;
+    padding: 50px;
+    justify-content: center;
+    border-style: dashed;
+    border-radius: 10px;
+    
+
+}
+
+.text1{
+    position: relative;
+    font-size: 20px;
+    margin: 0;
+    position: absolute;
+    top: 95%;
+    left: 50%;
+    margin-right: -50%;
+    transform: translate(-50%, -50%);
+    
+}
+.text2{
+    position: relative;
+    font-size: 20px;
+    margin: 0;
+    position: absolute;
+    top: 95%;
+    left: 80%;
+    margin-right: -50%;
+    transform: translate(-50%, -50%)
+}
+
+
+
